fix glass

+++ src/base/light.h
@@ -31,7 +31,7 @@ public:
     struct Evaluation {
         SampledSpectrum L;
         Float pdf;
-        [[nodiscard]] static auto zero(size_t spec_dim) noexcept {
+        [[nodiscard]] static auto zero(uint spec_dim) noexcept {
             return Evaluation{.L = SampledSpectrum{spec_dim}, .pdf = 0.f};
         }
     };
@@ -40,7 +40,7 @@ public:
         Evaluation eval;
         Float3 wi;
         Float distance;
-        [[nodiscard]] static auto zero(size_t spec_dim) noexcept {
+        [[nodiscard]] static auto zero(uint spec_dim) noexcept {
             return Sample{.eval = Evaluation::zero(spec_dim),
                           .wi = make_float3(0.f, 0.f, 1.f),
                           .distance = 0.f};
